Extract delete code to CuratorTestHelpers, use in ManagedLeaderLatchTest
* Extract the path deletion code to CuratorTestHelpers and use in
  both ManagedLeaderLatchCreatorTest and ManagedLeaderLatchTest
* Fix WhenWhen in test names in ManagedLeaderLatchTest
* Add assertion in (renamed) shouldCallActionAsynchronously_WhenIsLeader
  in ManagedLeaderLatchTest to ensure leaderLatch1 is leader before even
  attempting to run the test
